requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.22

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0
pdfplumber>=0.10.0

# Prompt compression
llmlingua>=0.2.0

# Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# OpenAI
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0

# Type hints
pydantic>=2.5.0
